# chore(sql): simpify offset calculation in varchar null check JIT code