… core/src/main/resources/io/questdb/bin/freebsd/libquestdb.so
… core/src/main/resources/io/questdb/bin/linux/libquestdb.so
… core/src/main/resources/io/questdb/bin/osx/libquestdb.dylib
… core/src/main/resources/io/questdb/bin/windows/libquestdb.dll