Minor code simplification.

Before:
```
Benchmark                                                  (a_nullFreq)  (b_useComplexFilter)  (c_jitMode)  Mode  Cnt   Score   Error  Units
StringNullCheckBenchmark.selectCountWhereVarcharColIsNull             0                 false         SIMD  avgt    5  12.907 ± 1.750  ms/op
StringNullCheckBenchmark.selectCountWhereVarcharColIsNull             0                  true         SIMD  avgt    5  17.344 ± 3.641  ms/op
StringNullCheckBenchmark.selectCountWhereVarcharColIsNull             1                 false         SIMD  avgt    5  33.954 ± 1.154  ms/op
StringNullCheckBenchmark.selectCountWhereVarcharColIsNull             1                  true         SIMD  avgt    5  44.665 ± 4.612  ms/op
StringNullCheckBenchmark.selectCountWhereVarcharColIsNull             3                 false         SIMD  avgt    5  30.702 ± 1.353  ms/op
StringNullCheckBenchmark.selectCountWhereVarcharColIsNull             3                  true         SIMD  avgt    5  40.807 ± 1.125  ms/op
StringNullCheckBenchmark.selectCountWhereVarcharColIsNull           100                 false         SIMD  avgt    5  14.635 ± 1.919  ms/op
StringNullCheckBenchmark.selectCountWhereVarcharColIsNull           100                  true         SIMD  avgt    5  19.344 ± 2.613  ms/op
```

After:
```
Benchmark                                                  (a_nullFreq)  (b_useComplexFilter)  (c_jitMode)  Mode  Cnt   Score   Error  Units
StringNullCheckBenchmark.selectCountWhereVarcharColIsNull             0                 false         SIMD  avgt    5  12.508 ± 0.200  ms/op
StringNullCheckBenchmark.selectCountWhereVarcharColIsNull             0                  true         SIMD  avgt    5  16.134 ± 0.868  ms/op
StringNullCheckBenchmark.selectCountWhereVarcharColIsNull             1                 false         SIMD  avgt    5  32.707 ± 0.582  ms/op
StringNullCheckBenchmark.selectCountWhereVarcharColIsNull             1                  true         SIMD  avgt    5  44.945 ± 1.404  ms/op
StringNullCheckBenchmark.selectCountWhereVarcharColIsNull             3                 false         SIMD  avgt    5  30.255 ± 0.970  ms/op
StringNullCheckBenchmark.selectCountWhereVarcharColIsNull             3                  true         SIMD  avgt    5  40.913 ± 1.030  ms/op
StringNullCheckBenchmark.selectCountWhereVarcharColIsNull           100                 false         SIMD  avgt    5  14.267 ± 1.125  ms/op
StringNullCheckBenchmark.selectCountWhereVarcharColIsNull           100                  true         SIMD  avgt    5  19.040 ± 0.454  ms/op
```